LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audiobuffersourcenode-interface/audiobuffersource-playbackrate-zero-expected.txt

CONSOLE MESSAGE: SyntaxError: The string did not match the expected pattern.

Harness Error (FAIL), message = SyntaxError: The string did not match the expected pattern.
